chore(mssql): remove legacy antlr parser dependency
