+++ backend/plugin/schema/mssql/register.go
@@ -1,0 +1,15 @@
+package mssql
+
+import (
+	storepb "github.com/bytebase/bytebase/backend/generated-go/store"
+	"github.com/bytebase/bytebase/backend/plugin/schema"
+)
+
+func init() {
+	schema.RegisterGetDatabaseMetadata(storepb.Engine_MSSQL, GetDatabaseMetadata)
+}
+
+type tableKey struct {
+	schema string
+	table  string
+}
